+++ src/tUilKit/__init__.py
@@ -3,6 +3,7 @@ from tUilKit.factories import (
     get_colour_manager,
     get_file_system,
     get_config_loader,
+    get_cli_menu_handler,
     reset_factories,
 )
 
@@ -13,5 +14,6 @@ __all__ = [
     "get_colour_manager",
     "get_file_system",
     "get_config_loader",
+    "get_cli_menu_handler",
     "reset_factories",
 ]